Major changes, added zOperands class. See description
Added "zemaxOperands.py" which contains the zemaxOperands module. This
module contains the zOperands class. Essentially, it has 3 class
variables all of which are dictionaries of zemax 4-character operands
(keys) and short description (value). They enable multiple use,
including but not limited to validating the correctness of input
operands in some functions, quickly listing the operands and short
descriptions, etc. Also added functions for getting and setting
tolerance in the tolerance editor.