Added ipCaptureWindow2, & small code updates
Added ipCaptureWindow2() for capturing analysis window from Zemax main
window and displaying on IPython Webnotebook or QtConsole. This function
is different from ipCaptureWindow as it passed analysis type code
instead of window number, and uses the metafile export from Zemax as the
source image. So the quality is generally better.