Bug fix: shutdown DDE server if ZMX not running
If zDDEInit() is called without a running Zemax application, subsequent
initialization used to fail even if zemax was started. This is because
the DDE server was not being shutdown in that scenario. One would have
to close the application and environment completely to use PyZDDE again.

This commit fixes that issue. If zDDEInit() is called without Zemax
running, an error message is printed to notify the user that Zemax may
not be running. Now, the user can run zemax and try zDDEInit() again
without closing the application.