Bug Fix: ipzCaptureWindow2() for earlier ZEMAX
Issue # 1: ipzCaptureWindow2() stalled.
The function ipzCaptureWindow2() was using the default lens path for
extracting, processing and then deleting the temporary metafile.
However, in earlier versions of ZEMAX (sometimes prior to approx. 2009)
this path is in `C:\Program Files\Zemax\Samples`. Accessing this folder
is likely not work due to permission issues. After the fix, this
function is going to use the directory of the lens file (not the default
lens path as set in the preferences of Zemax).