Minor updates
Added __eq__ to the PyZDDE class.
PyZDDE object gets an app name when initialized (pyz.PyZDDE()) and
doesn't have to wait for zDDEInit() to be successfully called. This
change was made to create consistency between the two methods ---
pyz.createLink() and pyz.PyZDDE() + pyz.zDDEInit() --- of creating DDE
link objects