fixed assumption in function zGetHiatus()
The function zGetHiatus() assumed that the first surface is a refracting
surface. Although, this assumption is generally true for 90% of cases,
it is not absolutely certain in all systems. In systems for which the
first surface is not a refracting surface, the hiatus calculation will
end up being wrong as the ENPP and other cardinal planes of the object
space are reported with respect to the first surface.