changed global variable APR flag to instance variable .apr
Based on usage, I have found that making APR a global switch to perform
automatic push and refresh can be problematic if working simultaneously
with two instances of PyZDDE objects -- i.e. the global variable will be
True for both. Therefore, I am changing this switch to be an instance
variable (property) which can be turned on or off any time for any
instance. The switch can also be turned on during object instantiation.